+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Version 2022-dev
 -  made membervariable format consistent (#694)
 -  add ability to run ibi as a postupd method (#696)
 -  removed unused functions (#702)
+-  reworked lammps molecule naming (#703)
 
 Version 2021.2 (released XX.07.21)
 ==================================